Added files for v1.0
This upload reflects the status of the code as of 2024/10/08.
It has been uploaded in preparation for the departmental presentation, held 24/10/11.
The code works as is, but is not user friendly as of yet. A comprehensive overhaul and repackaging will be performed prior to final publication, in order to facilitate replication.